created model module
